Fixed a large number of warnings when compiled under C++23 with -Wall - C++20 deprecates implicit enum to float conversion and enum/enum arithmetic - Some variables that appeared to be used uninitialized were explicitly initialized to 0/NULL - A few member initializers were out of order, they were put in class order - some incorrect sprintf format specifiers were fixed - some unsigned/int comparisons were coerced to int/int - Switch [[fallthrough]] is used if case fall-through is desired. This uses a new   AGG_FALLTHROUGH symbol

+++ examples/compositing.cpp
@@ -238,7 +238,7 @@ public:
 
         rb.blend_from(prim_pixfmt_type(rbuf_img(1)), 
                       0, 250, 180, 
-                      agg::cover_type(m_alpha_dst.value() * agg::cover_full));
+                      agg::cover_type(m_alpha_dst.value() * (double)agg::cover_full));
 
         circle(rb, 
                agg::srgba8(0xFD, 0xF0, 0x6F, unsigned(m_alpha_dst.value() * 255)),
